+++ ESI.NET/Http/EsiHeadersHandler.cs
@@ -19,6 +19,8 @@ namespace ESI.NET.Http
 
         public EsiHeadersHandler(IOptions<EsiConfig> config)
         {
+            if (config == null) throw new ArgumentNullException(nameof(config));
+
             _userAgent = config.Value?.UserAgent;
             if (string.IsNullOrWhiteSpace(_userAgent))
                 throw new ArgumentException(
@@ -28,6 +30,8 @@ namespace ESI.NET.Http
 
         protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
         {
+            if (request == null) throw new ArgumentNullException(nameof(request));
+
             if (!request.Headers.Contains("X-User-Agent"))
                 request.Headers.Add("X-User-Agent", _userAgent);