+++ ESI.NET/Http/EsiTokenRefreshHandler.cs
@@ -26,12 +26,16 @@ namespace ESI.NET.Http
 
         public EsiTokenRefreshHandler(IOptions<EsiConfig> config, IServiceScopeFactory scopeFactory = null)
         {
+            if (config == null) throw new ArgumentNullException(nameof(config));
+
             _config = config.Value;
             _scopeFactory = scopeFactory;
         }
 
         protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
         {
+            if (request == null) throw new ArgumentNullException(nameof(request));
+
             var character = EsiRequestState.GetCharacter(request);
 
             if (character != null